+++ src/proxy/residue-proxy.js
@@ -78,6 +78,9 @@ ResidueProxy.prototype = {
     get chainname () {
         return this.chainStore.getChainname( this.chainIndex );
     },
+    get chainid () {
+        return this.chainStore.getChainid( this.chainIndex );
+    },
 
     //