+++ src/store/chain-store.js
@@ -28,7 +28,8 @@ ChainStore.prototype = Object.assign( Object.create(
         [ "residueOffset", 1, "uint32" ],
         [ "residueCount", 1, "uint32" ],
 
-        [ "chainname", 4, "uint8" ]
+        [ "chainname", 4, "uint8" ],
+        [ "chainid", 4, "uint8" ]
 
     ],
 
@@ -54,6 +55,31 @@ ChainStore.prototype = Object.assign( Object.create(
             }
         }
         return chainname;
+
+    },
+
+    setChainid: function( i, str ){
+
+        var j = 4 * i;
+        this.chainid[ j ] = str.charCodeAt( 0 );
+        this.chainid[ j + 1 ] = str.charCodeAt( 1 );
+        this.chainid[ j + 2 ] = str.charCodeAt( 2 );
+        this.chainid[ j + 3 ] = str.charCodeAt( 3 );
+
+    },
+
+    getChainid: function( i ){
+
+        var chainid = "";
+        for( var k = 0; k < 4; ++k ){
+            var code = this.chainid[ 4 * i + k ];
+            if( code ){
+                chainid += String.fromCharCode( code );
+            }else{
+                break;
+            }
+        }
+        return chainid;
 
     }